+++ fase-01/camadas/restaurantes/src/main/java/com/postech/restaurantes/vo/v1/response/RoleResponse.java
@@ -1,10 +1,12 @@
 package com.postech.restaurantes.vo.v1.response;
 
+import java.util.UUID;
+
 /**
  * VO de saída para papel/role (v1).
  */
 public record RoleResponse(
-        Long id,
+        UUID id,
         String name
 ) {
 }